GeometricBase

Documentation for GeometricBase.

GeometricBase.AbstractScalarVariable

GeometricBase.AbstractSolverState

GeometricBase.AbstractStateVariable

GeometricBase.AbstractVariable

GeometricBase.GeometricData

GeometricBase.NullInvariants

GeometricBase.NullParameters

GeometricBase.NullPeriodicity

GeometricBase.NullSolverState

GeometricBase.State

GeometricBase.StateVariable

GeometricBase.StateVector

Base.copy!

Base.copy!

Base.getindex

Base.keys

GeometricBase.HistoryState

GeometricBase.SolverState

GeometricBase.integrate

GeometricBase.integrate!

GeometricBase.vectorfield

GeometricBase.zerovector

Library

GeometricBase.AbstractScalarVariable — Type

AbstractScalarVariable{T} is a wrapper around a zero-dimensional AbstractArray{T,0} that provides context for the nature of the variable, e.g., time.

GeometricBase.AbstractSolverState — Type

AbstractSolverState

The state is used in solvers and optimizers to translate information about previous solver/optimization steps to successive iterations.

GeometricBase.AbstractStateVariable — Type

AbstractStateVariable{T,N,AT} is a wrapper around a AT <: AbstractArray{T,N} that provides context for the nature of the variable, e.g., a state or a vector field.

GeometricBase.AbstractVariable — Type

AbstractVariable{T,N} is a wrapper around a AbstractArray{T,N} that provides context for the nature of the variable.

GeometricBase.GeometricData — Type

GeometricData is a wrapper for different types of data sets such as the solution of a Lagrangian or Hamiltonian system.

GeometricBase.NullInvariants — Type

NullInvariants is an empty struct that is used to indicate than an equation does not have invariants.

GeometricBase.NullParameters — Type

NullParameters is an empty struct that is used to indicate than an equation does not have parameters.

GeometricBase.NullPeriodicity — Type

NullPeriodicity is an empty struct that is used to indicate than an equation does not have periodic solution components.

GeometricBase.NullSolverState — Type

NullSolverState is an empty struct that is used to indicate than a solver does not store a state.

GeometricBase.State — Type

Holds the solution of a geometric equation at a single time step.

It stores all the information that is required to uniquely determine the state of a systen, in particular all state variables and their corresponding vector fields.

GeometricBase.StateVariable — Type

StateVariable{T,N,AT,RT,PT} is a wrapper around a AT <: AbstractArray{T,N} that holds one of the variables consituting the state of a dynamical system.

The value field holds the actual data, the range field holds the range of valid values, and the periodic field holds a bitmask indicating which dimensions are periodic.

GeometricBase.StateVector — Type

StateVector{DT,VT} is a vector of StateVariables, where DT is the datatype of the state and VT is the type of the vector.

Base.copy! — Method

initialize!(st::State, ics::NamedTuple)

Copy the values from a NamedTuple ics to the State st.

The keys of ics must be the same as the solution keys of the state.

Arguments

st: the state to copy into

ics: the named tuple containing the initial values to copy

Base.copy! — Method

copy!(dst::State, src::State)

Copy the values from one State src to another State dst.

The keys of src and dst must identical.

Arguments

src: the state to copy into

dst: the state containing the solution values to copy

Base.getindex — Method

getindex(st::State, args...)

Passes getindex on to the state in State.

Base.keys — Method

keys(st::State)

Return the keys of all the state variables in the State.

GeometricBase.HistoryState — Method

HistoryState(st::State)

Constructs a state whose symbols are decorated by a bar to indicate a previous value of the state.

GeometricBase.SolverState — Method

SolverState(::Union{<:AbstractSolver,<:SolverMethod}, args...; kwargs...)

This method returns a subtype of AbstractSolverState according to the AbstractSolver or SolverMethod that is passed to it. By default, i.e., if no such method is defined for a given solver, it returns NullSolverState.

GeometricBase.integrate — Function

integrate(problem, method; kwargs...)

Integrate a problem with method and return the solution.

GeometricBase.integrate! — Function

Solve one time step:

integrate!(integrator)

GeometricBase.vectorfield — Method

The vectorfield function returns a datastructure that stores the vectorfield for an AbstractStateVariable s that holds the state of a system. By default it returns a VectorfieldVariable that wraps zero(parent(s)), i.e., the same kind of array as the state variable, but custom methods can be implemented in order to account for more specific use cases, e.g., when s also contains constant fields that should not be present in the vector field.

GeometricBase.zerovector — Method

zerovector(X::StateVector) returns a new StateVector with zero applied all elements of X.