GeometricBase

Documentation for GeometricBase.

GeometricBase.AbstractScalarVariable

GeometricBase.AbstractStateVariable

GeometricBase.AbstractVariable

GeometricBase.GeometricData

GeometricBase.NullInvariants

GeometricBase.NullParameters

GeometricBase.NullPeriodicity

GeometricBase.StateVector

GeometricBase.integrate

GeometricBase.integrate!

GeometricBase.vectorfield

GeometricBase.zerovector

Library

GeometricBase.AbstractScalarVariable — Type

AbstractScalarVariable{T} is a wrapper around a zero-dimensional AbstractArray{T,0} that provides context for the nature of the variable, e.g., time.

GeometricBase.AbstractStateVariable — Type

AbstractStateVariable{T,N,AT} is a wrapper around a AT <: AbstractArray{T,N} that provides context for the nature of the variable, e.g., a state or a vector field.

GeometricBase.AbstractVariable — Type

AbstractVariable{T,N} is a wrapper around a AbstractArray{T,N} that provides context for the nature of the variable.

GeometricBase.GeometricData — Type

GeometricData is a wrapper for different types of data sets such as the solution of a Lagrangian or Hamiltonian system.

GeometricBase.NullInvariants — Type

NullInvariants is an empty struct that is used to indicate than an equation does not have invariants.

GeometricBase.NullParameters — Type

NullParameters is an empty struct that is used to indicate than an equation does not have parameters.

GeometricBase.NullPeriodicity — Type

NullPeriodicity is an empty struct that is used to indicate than an equation does not have periodic solution components.

GeometricBase.StateVector — Type

StateVector{DT,VT} is a vector of StateVariables, where DT is the datatype of the state and VT is the type of the vector.

GeometricBase.integrate — Function

integrate(problem, method; kwargs...)

Integrate a problem with method and return the solution.

GeometricBase.integrate! — Function

Solve one time step:

integrate!(integrator)

GeometricBase.vectorfield — Method

The vectorfield function returns a datastructure that stores the vectorfield for an AbstractStateVariable s that holds the state of a system. By default it returns a VectorfieldVariable that wraps zero(parent(s)), i.e., the same kind of array as the state variable, but custom methods can be implemented in order to account for more specific use cases, e.g., when s also contains constant fields that should not be present in the vector field.

GeometricBase.zerovector — Method

zerovector(X::StateVector) returns a new StateVector with zero applied all elements of X.